// src/app/Shared/Models/RequstService.ts
import { Time } from "@angular/common";

export class RequestService {
    id: number;
    serviceRequestCode: string;
    serviceRequestTitle: string;
    problemDescription: string;
    employeeEmail: string;
    serviceRequestTitleAr: string;
    problemDescriptionAr: string;
    serviceRequestDate: Date;
    serviceRequestTime: string;
    employeeId: string;
    requestType: string;
    EquipmentId: number;
    healthDirectoryId:number;
    healthDistrictId:number;
    modeId:number;
    priorityId:number;
}